pyproject.toml
[tool.poetry]
name = "getwvkeysbot"
version = "1.2.0"
description = "Discord Bot for GetWVKeys"
authors = ["Puyodead1 <puyodead@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
"discord.py" = { git = "https://github.com/Rapptz/discord.py" }
coloredlogs = "^15.0.1"
requests = "^2.28.0"
python-dotenv = "^0.20.0"
websockets = "^12.0"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
isort = "^5.10.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = 'getwvkeysbot.main:main'
